Changed the_full_path_to_some_empty_file to output_message_file.
When I first read this README I was wondering if the file had to exist
so that the okc_arrow_fetcher.py would write the messages in it or not.

So I did take a look at the source:

```
f = codecs.open(file_name, encoding='utf-8', mode='w')
```

This will just (over)write any file named 'file_name' regardless if it
exists already or not.