Add python virtual environment, strict package versions, wrapper script
Used to lock usage to Python 2.7 and BeautifulSoup 3.2.1. This is to reduce version confusion in error reports,
and to be able to keep track of dependency version upgrades. The wrapper script `okcmd` activates the virtual
environment in `.okcmd-virtualenv/` as well as installs dependencies -- usage is the same as for
`okc_arrow_fetcher.py`.

Requires `python2.7`, `virtualenv` and `pip` in the `$PATH`.