Alternative login method "autologin link"
Using the normal username/password combination when to log in to OkCupid might
not work. One theory is that it triggers automatic bot blocking which requires
a captcha to be filled in.

By instead using special links that OkCupid recognizes as secure enough to
automatically log you in, this bot protection can be bypassed. Auto-login
links can be found in emails sent out by OkCupid. One example is the OkCupid
logotype in new/top matches. Copy that link directly from your email, and
use it for the `--autologin` parameter, which then replaces `--password`.