Upgrade to BeautifulSoup 4
This should fix problems with tricky/malformed HTML being served on the messages page (<div> and other tags inside an <ul> tag). BS3 had problems parsing it and because of that the (around) three most recent messages in conversations which had more than three messages were not saved to disk properly.