Get message thread id (url) from data property
OKCupid have redesigned the messages/thread/conversation list, and removed links to the page displaying the content of each individual message. Instead messages are loaded in the medium size message box in the lower right corner of the screen, using ajax.

The old message display page (the one harvested by OKCMD) still exists on the server, but perhaps not for long. OKCMD users are encouraged to backup their messages ASAP, in case further design changes are harder to adapt to.