Save generated mutual match without messages
When there's a mutual match, but no actual messages have been sent, a dummy thread/conversation is created. This type of thread/conversation is listed among the other messages, and therefore OKCMD downloads it. The parser could previously not find any messages as it's displayed with a custom layout, but this has now been added to the output as a separate placeholder message.

While this is good for threads with no other messages, if a mutual_match_no_messages thread later has messages the layout changes. This will affect the diff between OKCMD runs. Perhaps a more stable solution would save additional metadata about each thread? As threads are now per-contact, this probably means saving user information.